60hz on real hardware
there isn't enough time to copy the entire screen using latches since
the vram read latency is too high.

instead, draw the bars into a buffer in ram and copy it to vram at each
row, so there are only vram writes.

but that's _still_ not enough to reach 60hz.

the ram row buffer is divided into planes, so that the vram can be
written with dword copies.  easy 60hz.